Xunfei speech recognition function added
Click on the "Speak" button in form2 (query) and start speaking, and
your speech will be recorded. When finished speaking, click on "Stop"
and the speech will be uploaded to Xunfei Cloud and translated to text
in the textbox. Both English and Chinese can be recognized by selecting
different radio buttons. Note that it might take a while to recognize
the speech if the sentence is too long.
Also the program can only record for about 10 seconds and it will crash.
I am working to fix the bug (hopefully when we are testing it we won't
talk for that long to the agent?)