Updated different language modes for display and tts
Used a different way to parse results. Radio buttons "English" and
"Chinese" in the query window still controls the language of what's
being displayed to users. And they can now use the "Same Lang." checkbox
to set the TTS language the same as or different than the language being
displayed. "TTS Enabled" checkbox let users choose to do TTS or not.
Also changed the language_mode in Constant.cs to be
language_mode_display and language_mode_tts