XML filer changes and bugfixes
Changed XML reading, writing, and merging functions in
XMLFilerForFeatureGraph to use same xml structure as front-end, with
Features, Neighbors, and Parents node containers. Fixed bug in lead-in
statements so it uses the correct previous feature instead of the
feature two statements back. Fixed bug in file opening where feature
graph and query handler would not reset and update to the feature graph
for the new file.